inventory/requirements.txt
django>2
djangorestframework==3.8.0
django-import-export
